This function uses rootfinding with fitted curve to compute IC values.

This function uses rootfinding with fitted curve to compute IC values.

computeIC(drFit, percent = 0.5, log.d = TRUE, interpolation = TRUE,
  stepLen = NA, lower, upper, ...)

Arguments

drFit
A drFit object as returned by drFit() function.

percent
the inhibition ratio to be searched against. A vector between 0 and 1. Corresponding response is 1 - percent

log.d
whether to return log10(dose) or the raw dose. Default is set to TRUE.

interpolation
whether to use interpolation to estimate IC values. In this case, the computed IC values will be bound by the observed dosages.

stepLen
step length to construct equally spaced intervals during interpolation. Only used when interpolation = TRUE.

lower
lower bound in root search. Default is min(c(0, min(dose))) where dose is the observed dose levels from drFit@originalDat.

upper
upper bound in root search. Default is max(dose)*1e6 where dose is the observed dose levels from drFit@originalDat.

...
other optimization parameters to be passed to uniroot().

Value

A named vector giving the IC values at specified percentiles.

See also

Examples

data(ryegrass, package = 'drc') # use the ryegrass data from drc package
fit.sigEmax <- drFit(drMat=ryegrass[, c(2, 1)], modelName = "sigEmax", alpha = 0.01, fitCtr = FALSE)

#> Message: Need bounds in "bnds" for nonlinear models, using default bounds from "defBnds".

computeIC(fit.sigEmax, percent = seq(0, 1, by = 0.1), log.d = FALSE)

#>       IC0      IC10      IC20      IC30      IC40      IC50      IC60      IC70 
#>  0.940000  1.538414  1.967747  2.363281  2.751399  3.192205  3.716456  4.435611 
#>      IC80      IC90     IC100 
#>  5.615728  9.064101 30.000000