Implementation of NCI60 method (GI50, TGI, LC50) for dose response data

This function implements S3-style OOP. Methods such as predict, plot and lines are available. Notice that here the response is assumed to be calculated as in: http://dtp.nci.nih.gov/branches/btb/ivclsp.html. The response has range -1 to 1.

nci60Fit(d, y, interpolation = TRUE, log.d = FALSE)

Arguments

d
dose original dose.

y
response percentage of growth inhibition as defined by NCI60 method.

interpolation
whether to use interpolation to estimat GI50, TGI and LC50.

log.d
whether to return log10 transformed value for GI50, TGI and LC50 (the dose concentration).

Value

a nci60Fit object consisting different attributes. It includes a vector with GI50, TGI and LC50.

Details

The original NCI60 method applies linear interpolation to estimate GI50, TGI and LC50. Tong, F. P. improved this by using 4-parameter logistic model, which is the LL.4 model in drexplorer (from drc package). For resistant cell lines, even the LL.4 cannot be fitted since the curve is essentially flat. Here we fit multiple models throught the model selection framework in drexplorer and use the best model to estimate GI50, TGI and LC50. This will almost give an estimate for all extreme data without giving NA.

References

Shoemaker, R. H. The NCI60 Human Tumour Cell line Anticancer Drug Screen. Nature Reviews, 6: 813-823, 2006.

Tong, F. P. (2010). Statistical Methods for Dose-Response Assays.

Examples

fit_nci60 <- nci60Fit(d = datNCI60$Dose, y = datNCI60$Growth/100)

#> Message: Need bounds in "bnds" for nonlinear models, using default bounds from "defBnds".

#> Message: Need bounds in "bnds" for nonlinear models, using default bounds from "defBnds".

fit_nci60[1:length(fit_nci60)]

#>       GI50        TGI       LC50     AOC_GR     AOC_LD    AUC_GRI 
#> 0.06825010 0.27381347 0.79000000 0.42983275 0.02088332 0.77464196