plot method for drFit object

plot method for drFit object

plot(x, y, ...)

# S4 method for drFit
plot(x, pchs = c(16, 17, 15), cols = c(1, 2, 3),
  col = 4, lwd = 2, lty = 1, addLegend = FALSE, xlab = "Dose",
  ylab = "Relative viability", xlim = NA, ylim = NA, main,
  type = c("line", "points", "control", "legend", "sem"), style = "",
  bty = "n", h = c(0.5), cex.main = 1, cex.axis = 1, cex.lab = 1,
  axes = TRUE, return = FALSE)

Arguments

x
a drFit object.

y
y is not required in a drFit object.

...
Arguments to be passed to methods, such as graphical parameters (see par).

pchs
pchs (a vector) specify symbols to show different data points. In particular, pchs[1] is used to show regular points. pchs[2] is used to show outliers at 0.05 significance level and pchs[3] is used to show outliers at 0.01 significance level.

cols
Similarly, cols[1] is used to color regular points, cols[2] to color outliers at 0.05 significance level and cols[3] to color outliers at 0.01 significance level.

col
color used to specify the appearance of fitted curve.

lwd
line width used to specify the appearance of fitted curve.

lty
line type used to specify the appearance of fitted curve.

addLegend
currently deprecated. Please specify through type argument. Whether to add legend indicating outlier status.

xlab
x axis label.

ylab
y axis label.

xlim
y limit for display.

ylim
x limit for display.

main
main title.

type
the plot type. The user can specify more flexibly by specifying a vector of entities to plot including: line: dose response curve points: observed data points control: control data points legend: legend for outlier status

style
deprecated. Please specify through type argument. when style='full', observed dose-response pairs are plotted including controls, fitted curve is superimposed and legend for outliers indicated; when style='simple', only fitted curve is plotted (without dose-response points and of course, not outlier status).

bty
bty passed to legend.

h
horizontal line to add indicating e.g. IC50 (h=0.5).

cex.main
cex for main title.

cex.axis
cex for axis annotation.

cex.lab
cex for axis label.

axes
whether to add axes.

return
whether return data for the plotted lines; this is useful if we want to reconstruct the lines elsewhere.