Estimate the interaction index as well as its confidence interval using delta method for fixed ratio design

This code is extracted from the source code distributed at https://biostatistics.mdanderson.org/SoftwareDownload/

fitIAI(d1, d2, e, E = seq(0.05, 0.95, 0.005), name1 = "Drug A",
  name2 = "Drug B", alpha = 0.05, d2.d1.force = NA, tol = 0.01)

Arguments

d1
dose for drug 1.

d2
dose for drug 2.

e
corresponding response in the range [0, 1].

E
a vector of responses (between 0 and 1) where IAI and confidence interval are to be computed from.

name1
name of drug 1.

name2
name of drug 2.

alpha
significance level of confidence interval.

d2.d1.force
a ratio passed to detect_ray_design() function so as to specify a fixed ratio for grid design.

tol
tolerance in declaring fixed ratio.

Value

a data frame with columns IAI, IAI.low, IAI.up, E, dx1 (corresponding dose of drug 1), dx2 (corresponding dose of drug 1).

Details

Two papers have been published by Lee et al, one in 2007 (Lee2007) and on in 2009 (Lee2009). The Lee2007 paper described five methods to assess interaction: (1) Lowewe additivity model using interaction index (IAI) (2) Model of Greco et al 1990. This approach uses $$\alpha$$ as the metric and it can be related to IAI (3) Model of Machado and Robinson which uses a metric denoted as $$\eta$$ (4) Model of Plummer and Short which can also be linked to IAI through the parameter $$\beta_4$$ (5) Model of Carter et al that can be linked to IAI through the parameter $$\beta_{12}$$. For more details of these models, please refer to Lee2007.

The Lee2009 paper provided generalization of IAI to multiple drugs using Lowewe additivity model and assumption of Chou and Talalay's median effect equation. The Chou and Talalay's median effect equation can be expressed as: $$log(E/(1-E))=m(log d - log Dm)$$ where E is the effect at dose d for a compound whose median effective dose.

Some notes about experiment design. Usually the data is either fixed ratio design (ray design) or grid design which means all-possible combination of drug concentrations between two drugs are available. The Lee2007 paper provided an example of grid design. However, specific fixed ratio is used to fit the median effect model which is the basis to estimate IAI. The Lee2009 paper considered with fixed ratio design.

References

Lee, J. J., & Kong, M. (2009). Confidence intervals of interaction index for assessing multiple drug interaction. Statistics in biopharmaceutical research, 1(1), 4-17.

Lee, J. Jack, et al (2007). Interaction index and different methods for determining drug interaction in combination therapy. Journal of biopharmaceutical statistics 17.3 461-480.

Examples

fit_allPoss_1 <- fitIAI(d1 = nl22B2$schd, d2 = nl22B2$hpr, e = nl22B2$y1,
                        name1 = 'SCH66336', name2 = '4HPR', d2.d1.force = 1)

#> Warning: This is not a ray design; computation is done with method derived from fixed ratio!